Isolate test failure, simplify
